Fix bug where empty slice of blocks doesn't update module
The fix was just not sending back an empty slice of blocks when the
module doesn't have any contextual blocks to send. In the case of the
network module, this is sending a single block that just reads that the
system is not connected via any interfaces (i.e. an error message).